Use a different set of conditions for suggesting Purifying Brew.
Instead of suggesting PB whenever Stagger is yellow/red (moderate/heavy),
calculate the remaining damage from the existing Stagger DoT and suggest
PB if the remaining damage is at least 30% of the monk's maximum health.

Also suggest PB if the Stagger DoT tick is more than half of the remaining
health.

The rationale is that a high damage DoT with only a second or two left
might not actually be dangerous to the monk's survival and the Chi could
be better spent on avoidance abilities or self-healing.