Update Nerien's frost mage script to SimC:sc_mage.cpp (r15073).
Refactor the sections of the SimC profile dealing with the Tier 6 talents
so that the common sections and logic is more clear.

Move the short cooldown icon in front of the main action icon.  This will
be the new standard icon placement for Nerien's scripts.

Fix issue #1, "Script not recognizing Brilliant Mana Gem", by adding
support for Glyph of Mana Gem.