Moving Healing Touch into main actions for the DoC rotation.
Healing Touch is a proc that must be consumed in fairly short order so
there very little point in ever delaying or wasting it.  Even if it can't
be used on Rip or Rake, it can be used to power any of the other direct
yellow attacks.

Rearrange icons so that short CDs are on the left side.