Monk: By default, pool energy as a brewmaster for survival abilities.
Add a new checkbox for "aggressive Jabbing", off by default, that has the
old behavior of suggesting Jab or SCK as long as there is enough energy.
The new default behavior is to pool energy and only Jab/SCK to prevent
capping energy.  This always leaves enough energy for using Expel Harm or
Healing Sphere on demand.