Be explicit on conditions for abilities only usable at certain times.
For abilities that are usable only if the target is below a certain health
percent threshold, list the exact health percent as part of the condition
for using the ability.  This allows Ovale to see ahead of time whether a
spell can be used instead of seeing if it's usable after the current GCD
is up.

The affected abilities are a rogue's Dispatch and a hunter's Kill Shot.