Druid: Initial balance script based on SimC:sc_druid.cpp (r16833).
This script supports the single-target rotation for balance druids
assuming a Patchwerk-style fight. It is up to the player to recognize
when a high-movement or AoE phase is coming and to work to stay in Eclipse
for that phase.

The four icons, left to right, are:

    short CD, main, moving, CD

where "moving" shows the rotation to be used while moving.

The following bits of a normal moonkin rotation are not yet supported:

    Wild Mushrooms
    Overwriting DoTs for trinket procs