Monk: Initial mistweaver script.
There isnt much to say for single-target healing -- mistweavers should
know how to do that on their own. The main icon instead shows an optimal
"fist-weaving" rotation that maintains uptime on the required buffs and
manages mana for that rotation.  An AoE rotation that reminds when to cast
Renewing Mist and Uplift is also included.