fix(deathknight): fixes for Night Fae blood death knights
Use the correct spell ID for the Death's Due buff, and remove the
'dnd_buff' symbol which was unnecessary since Night Fae death
knights still gain the normal Death and Decay buff from standing
in Death's Due.

Adjust the checks for the Death's Due buff/debuff to actually
check on both the player and the target.