feat(mage): preliminary script for arcane mage
Add preliminary support for arcane mages with a script based on
Toegrinder's Arcane Mage Compendium.

Currently only adapts to Necrolord and Venthyr covenant abilities.

The icon layout is:

  - Arcane Power Burn actions
  - Mini Burn actions
  - Conserve actions
  - AoE actions

There is no short CD or CD actions icon as with other classes due
to the complexity of the mana management for arcane mage.