fix(monk): suggest Invoke Niuzao more aggressively for brewmaster
Invoke Niuzao should be used in the opener for brewmaster monks
to reduce damage taken on a pull and for instant offense. It should
also be used whenever the pull should last at least 25 seconds
more, but that is difficult to gauge correctly, so only use that
condition for the primary target, otherwise suggest is always for
AoE and let the player decide to use that offensive CD or not.