fix(monk): pool energy for double Keg Smash
Pool energy for double Keg Smash if Weapons of Order is coming off
cooldown. We do this by slowly scaling up the perceived cost of Keg
Smash so that when Weapons of Order is off cooldown, it will equal
the cost of a double Keg Smash. This slowly pools the energy over
the entire cooldown period of Weapons of Order by slightly delaying
a energy-spending filler attack each time.