fix(paladin): adjust Seraphim and Shield of the Righteous usage
Move Seraphim to the short CD icon as it's a holy power spender,
and suggest Shield of the Righteous to avoid capping on holy
power. Remove the check for the Judgment debuff on the target
because it doesn't affect whether we plan on casting SotR or not.

Also suggest Word of Glory to heal others when it's free.