fix(warrior): move Last Stand usage to a heal action
Last Stand is more of a heal (for 30% more health) than a defensive
cooldown, which is usually for damage reduction or increased
healing.  Move Last Stand to the heal actions and simplify the
conditions for suggesting Shield Wall.

This should suggest Shield Wall more aggressively to put it on
cooldown, which works much better especially if Anger Management or
The Wall legendary is used.