fix(warrior): improve protection tier set bonus usage
When Seeing Red is at 7 stacks and about to transition to Outburst,
put Thunder Clap on cooldown and suggest the appropriate action for
proccing Outburst to ensure that Shield Slam is used to consume
Outburst.

Similarly, when Avatar is coming off cooldown with the 4-piece tier
bonus, put Thunder Clap on cooldown so that Shield Slam has the
highest priority when Avatar is cast and Outburst is procced.