fix: improve "cycle" suggestions for choosing a different target
Match the way that the assassination rogue script suggests applying
a DoT onto a different target by using the annotion "new" for
applying the DoT onto the current target that's missing the DoT,
the annoation "other" for applying the DoT onto a different target,
and no annotation for refreshing the current DoT on the target.
This makes it more explicit that a new target should be chosen for
the action instead of a generic "cycle" that relies on the player
to possibly find a new target.